+++ packages/react-instantsearch-router-nextjs/CHANGELOG.md
@@ -3,6 +3,22 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [7.21.0](https://github.com/algolia/instantsearch/compare/react-instantsearch-router-nextjs@7.20.2...react-instantsearch-router-nextjs@7.21.0) (2025-12-10)
+
+**Note:** Version bump only for package react-instantsearch-router-nextjs
+
+
+
+
+
+## [7.20.2](https://github.com/algolia/instantsearch/compare/react-instantsearch-router-nextjs@7.20.1...react-instantsearch-router-nextjs@7.20.2) (2025-12-09)
+
+**Note:** Version bump only for package react-instantsearch-router-nextjs
+
+
+
+
+
 ## [7.20.1](https://github.com/algolia/instantsearch/compare/react-instantsearch-router-nextjs@7.20.0...react-instantsearch-router-nextjs@7.20.1) (2025-12-02)
 
 **Note:** Version bump only for package react-instantsearch-router-nextjs